COH-594 Changes to transaction granularity
Reinstated transactional methods to EntityDatabase (begin/commit/abort)
so that we can give control over the granularity of transactions to
the calling code.

Also, added a close method to EntityDatabase. In the TDB implementation
this is used to free up any resources held by TDB's caches for a dataset

Change-Id: Id2d5691c525e47c0d072ff39f350f1068a8ceb39

Conflicts:

	pom.xml
	src/main/java/com/talis/entity/EntityDatabase.java
	src/main/java/com/talis/entity/db/ram/RamEntityDatabase.java
	src/main/java/com/talis/entity/db/tdb/TdbEntityDatabase.java
	src/test/java/com/talis/entity/db/tdb/TdbEntityDatabaseTest.java